Migration of ScheduleBuilder class to Schedule class
